[S09P12B207-292] feat : 식사, 수면 카테고리 일별 통계 기능 추가 - 윤민재

See merge request s09-webmobile2-sub2/S09P12B207!118

+++ backend-web/src/main/java/com/fa/sonagi/statistics/diaper/dto/DiaperStatisticsResDto.java
@@ -3,7 +3,6 @@ package com.fa.sonagi.statistics.diaper.dto;
 import java.util.ArrayList;
 import java.util.List;
 
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -12,12 +11,12 @@ import lombok.Setter;
 public class DiaperStatisticsResDto {
 	private List<DiaperStatisticsQueryDto> pees;
 	private List<DiaperStatisticsQueryDto> poops;
-	private Integer peeCnt;
-	private Integer poopCnt;
-	private Integer peePercent;
-	private Integer yesterdayPeePercent;
-	private Integer poopPercent;
-	private Integer yesterdayPoopPercent;
+	private Long peeCnt;
+	private Long poopCnt;
+	private Long peePercent;
+	private Long yesterdayPeePercent;
+	private Long poopPercent;
+	private Long yesterdayPoopPercent;
 
 	public DiaperStatisticsResDto() {
 		this.pees = new ArrayList<>();
